started handle cli options

+++ tasks.py
@@ -1,5 +1,6 @@
-from pathlib import Path
+import argparse
 import json
+from pathlib import Path
 
 class Task(object):
 
@@ -74,7 +75,12 @@ class Tasklist(object):
             self.num_of_tasks -= 1
             self.return_id(id_)
 
-    def change(self, id_, text):
+    def remove_all(self):
+        self.tasks.clear()
+        self.free_ids = [i for i in range(1, 100)]
+        self.num_of_tasks = 0
+
+    def edit(self, id_, text):
         if id_ in self.tasks:
             self.tasks[id_].set_task(text)
 
@@ -86,11 +92,6 @@ class Tasklist(object):
         temp_dict = self.tasks.copy()
         for k in temp_dict:
             self.tasks[k].set_status('finished')
-
-    def clear_all(self):
-        self.tasks.clear()
-        self.free_ids = [i for i in range(1, 100)]
-        self.num_of_tasks = 0
 
     def list_all(self):
         print('\n' + self.get_name().upper())
@@ -106,11 +107,7 @@ class Tasklist(object):
 
 
 def write_tasks(tasks, path):
-    if path.is_dir():
-        # TODO: raise custom exception
-        pass
-    else:
-        if path.exists():
+        if path.exists() and path.is_file():
             try:
                 with path.open('w') as f:
                     data = {}
@@ -124,44 +121,88 @@ def write_tasks(tasks, path):
             except IOError:
                 # TODO: raise custom exception
                 pass
+        else:
+            # TODO: raise custom exception
+            pass
 
 
 def read_tasks(tasks, path):
-    if path.is_dir():
-        # TODO: raise custom exception
-        pass
+    if path.exists() and path.is_file():
+        try:
+            with path.open('r') as f:
+                data = json.load(f)
+                for k in data:
+                    tasks.add(int(k), data[k]['text'], data[k]['status'])
+        except IOError:
+            # TODO: raise BadPath exception
+            pass
     else:
-        if path.exists():
-            try:
-                with path.open('r') as f:
-                    data = json.load(f)
-                    for k in data:
-                        tasks.add(int(k), data[k]['text'], data[k]['status'])
-            except IOError:
-                # TODO: raise custom exception
-                pass
+        # TODO: raise BadPath exception
+        pass
 
 
 def main():
     tasks = Tasklist()
     path = Path().resolve().joinpath(tasks.get_tasksdir(),
                                      tasks.get_name() + '.json')
-    id_ = tasks.get_free_id()
-    tasks.add(id_, 'Programming')
-    id_ = tasks.get_free_id()
-    tasks.add(id_, 'Math')
-    id_ = tasks.get_free_id()
-    tasks.add(id_, 'Jogging')
-    id_ = tasks.get_free_id()
-    tasks.add(id_, 'Reading')
-    id_ = tasks.get_free_id()
-    tasks.add(id_, 'Movie')
     read_tasks(tasks, path)
-    tasks.remove(5)
-    #tasks.finish(5)
-    write_tasks(tasks.get_tasks(), path)
-    tasks.list_all()
 
+    parser = argparse.ArgumentParser()
+
+    parser.add_argument('-a', '--add', dest='add', type=str,
+            help='add TASK to tasklist')
+
+    parser.add_argument('-e', '--edit', dest='edit', nargs=2,
+            help='edit TASK with specified id')
+
+    parser.add_argument('-f', '--finish', dest='finish', type=int,
+            help='mark TASK as finished')
+
+    parser.add_argument('-F', '--finish-all', dest='finish-all',
+            action='store_true', help='mark all TASKS as finished')
+
+    parser.add_argument('-r', '--remove', dest='remove', type=int,
+            help='remove TASK from tasklist')
+
+    parser.add_argument('-R', '--remove-all', dest='remove-all',
+            action='store_true', help='remove all TASKS from tasklist')
+
+    parser.add_argument('-u', '--unfinish', dest='unfinish', type=int,
+            help='mark TASK as unfinished')
+
+    parser.add_argument('-U', '--unfinish-all', dest='unfinish-all',
+            action='store_true', help='mark all TASKS as unfinished')
+
+    args = parser.parse_args()
+
+    if args.add:
+        tasks.add(args.add)
+        write_tasks(tasks.get_tasks(), path)
+    elif args.finish:
+        tasks.finish(args.finish)
+        write_tasks(tasks.get_tasks(), path)
+    elif args.finish_all:
+        tasks.finish_all()
+        write_tasks(tasks.get_tasks(), path)
+    elif args.edit:
+        tasks.edit(int(args.edit[0]), args.edit[1])
+        write_tasks(tasks.get_tasks(), path)
+    elif args.remove:
+        tasks.remove(args.remove)
+        write_tasks(tasks.get_tasks(), path)
+    elif args.remove_all:
+        tasks.remove_all()
+        write_tasks(tasks.get_tasks(), path)
+    elif args.unfinish:
+        tasks.unfinish(args.unfinish)
+        write_tasks(tasks.get_tasks(), path)
+    elif args.unfinish_all:
+        tasks.unfinish_all()
+        write_tasks(tasks.get_tasks(), path)
+    else:
+        tasks.list_all()
+
+    # TODO: finish parse options and make pretty output
 
 if __name__ == '__main__':
     main()
